add new components for payment page
